src/requirements.txt
click==8.2.1
openai==1.84.0
platformdirs==4.3.8
polars==1.30.0
python-dotenv==1.1.0
PyYAML==6.0.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
